requirements.txt
flask
# psycopg2
pandas
requests
PyPI
psycopg2-binary
